+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,16 @@
 # sit-onyx
 
+## 1.0.0-beta.4
+
+### Patch Changes
+
+- 6e14afd: fix(OnyxMobileNavButton): scroll on overflowing mobile flyout
+
+  The flyout of OnyxMobileNavButton now has a max height and is scrollable if too many nav/context items exist.
+
+  - app version inside the mobile flyout is not positioned absolute anymore and is a disabled list item
+  - fixed duplicate border in mobile context menu when multiple list items exist
+
 ## 1.0.0-beta.3
 
 ### Minor Changes